pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "GsJrys"
version = "2.0.0"
description = "Gscore 插件 - 每日运势 (极简重构版)"
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "Pillow>=9.0.0",
    "httpx>=0.24.0"
]